add 1 min export feature

+++ plugins/event_store/event_export/task/include/event_read_handler.h
@@ -49,7 +49,7 @@ struct EventReadRequest : public BaseRequest {
     std::string exportDir;
 
     // task type
-    int64_t taskType;
+    int16_t taskType;
 };
 
 struct ExportPeriodInfo {
